[FIX/#7] 1주차 과제 이동
