+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
+    xmlns:tools="http://schemas.android.com/tools" >
 
     <application
         android:name=".MyApp"
@@ -12,10 +12,10 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.DoSoptTemplate"
-        tools:targetApi="31">
+        tools:targetApi="31" >
         <activity
             android:name=".presentation.auth.LoginActivity"
-            android:exported="true">
+            android:exported="true" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -27,6 +27,9 @@
             android:exported="false" />
         <activity
             android:name=".presentation.main.MainActivity"
+            android:exported="false" />
+        <activity
+            android:name=".presentation.main.HomeActivity"
             android:exported="false" />
     </application>
